Now clearing off any character not between A..Z Reorganized the code
