+++ src/lib.rs
@@ -1,4 +1,4 @@
+pub mod ciphers;
 pub mod errors;
 pub mod playfair;
-
 mod structs;
